Rename IPPort to ServerAddress and add server ids
## Motivation / Description
uhashring's HashRing uses str() on the entities added to the
ring to generate the hashes and the corresponding shardmap.

To mimic bmemcached behavior we need to customize the server
ids we use, so we get exactly the same shard map as before.

In general, controlling the server id can lead to better
management of the pool. If you want, for example, to replace
a server with another, if the server id is kept constant
you avoid any shardmap changes at all.

## Changes introduced
* Rename IPPort to ServerAddress and also related methods
* Add server_id, so the HashRing behavior can be controlled
  easily with the pool configuration