RFC: Moving serialization to Executor
## Motivation / Description
Executor holds the serializer and does deserialization of values. This
is required because we try to minimize memory copying, so if we can
deserialize directly from the byte buffer of the connection, we save
creating and passing intermediate byte arrays up to be deserialized
later.

But serialization now happens on the meta-commands, two levels up.
Right now we were punching a hole from meta-commands to the executor
to encode values with the serializer, which works but isn't pretty.

This explores moving the serialization to the Executor as well.

Pros:
* Cleaner, proper layers

Cons:
* We can potentially serialize twice if some routers / executors send
  the same value to more than one pool destination.