Add ProbabilisticHotCache pool wrapper
## Motivation / Description
This adds a pool wrapper that can cache locally the hotest
cache keys to reduce cache traffic and increase throughput.

The mechanism to detect hot keys is probabilistic. We ask
Memcache to return if the key was accessed since the write
and how many seconds ago was the last read of the key.

`max_last_access_age_seconds` configures the max read age
for a key to be considered "hot". Sometimes a key that is
not hot will be accessed twice at the same time, by pure
random behavior. That is why additionally we have the
`probability_factor`. Only once every `probability_factor`
we will consider the key hot, so controlling this
probability you can make sure that only the keys that
have high-concurrency, that are accessed at the same
time many times will eventually be marked hot.

Once marked hot, the value will be stored in the local
cache for `cache_ttl`. Normally is best to use low
values, like 60s so the local cache is revalidated often.

This cache also allows specifying allowed pregixes, so
you can restrict the local caching to data that is not
very sensitive, that changes rarely but is high-thoughput.