Add touch operations errors w/ low ttl to write failure tracker
## Motivation / Description
Touch operations with a low TTL are usually meant as a way to cause a delete
in the future.
By default we will consider touch with < 50s as a delete, and thus track
errors for them in the write failure tracker.

This can be disabled setting `touch_ttl_to_consider_write_failure = None`
in the Executor.

## Changes introduced
- Track errors on touch operations with low TTL
- Add tests