Expose track write failures flag in meta-commands (#47)
## Motivation / Description
We want to make possible to disable the write failure tracker
for certain commands, so we are exposing the track_write_failures
as a flag in all meta commands.

The rationale for this is that we want the refill write errors
to not cause a write failure and subsequent invalidation.

If a cache is down, reads will cause a miss. This will make
the client read from the backend and attempt to repopulate the
cache. If that write fails, it is fine to ignore it. Recording
it for invalidation might lead to invalidation storms and performance
degradation. Only writes during create, update and delete operations,
when the DB is actually being changed should be tracked as failures.

## Changes introduced
- Make routers pass the track_write_failure flag to executors
- Make mete commands expose track_write_failure flag
- Make a high-level refill command that will issue an set with mode ADD
  and disable write failure tracking.